backend_agents/requirements.txt
# DataCue Phase 1 Requirements

fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pandas==2.2.3
python-dotenv==1.0.1
python-multipart==0.0.12
groq==0.11.0
google-generativeai==0.8.3
requests==2.32.3
